+++ jericho/enums/cluster_roles.py
@@ -2,6 +2,6 @@ from enum import Enum
 
 
 class ClusterRole(Enum):
-    MASTER = "master"
-    SLAVE = "slave"
+    SOURCE = "source"
+    REPLICA = "replica"
     DISABLED = "disabled"